Fix harmless warning in wxOSX with 10.7 SDK.
Use NSInteger instead of int in a declaration to avoid warnings about
declaration mismatch from g++ 4.2 under OS X 10.7.

git-svn-id: https://svn.wxwidgets.org/svn/wx/wxWidgets/trunk@69988 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775

+++ src/osx/cocoa/textctrl.mm
@@ -325,7 +325,7 @@ NSView* wxMacEditHelper::ms_viewCurrentlyEdited = nil;
 }
 
 - (NSArray *)control:(NSControl *)control textView:(NSTextView *)textView completions:(NSArray *)words
- forPartialWordRange:(NSRange)charRange indexOfSelectedItem:(int*)index
+ forPartialWordRange:(NSRange)charRange indexOfSelectedItem:(NSInteger*)index
 {
     NSMutableArray* matches = NULL;
